Write event data from mmap buffer into file
This is the code necessary to write data from the ring buffer used by
perf_event_open() into a file for later use.  One interesting annoyance
of note is that the inherit bit in the perf_event_attr structure CANNOT
be set if you intend to use the mmap interface.  For reasons that are
not entirely clear to your frustrated commit message author.